+++ backend/src/main/java/com/example/crypto_trading_sim/models/Account.java
@@ -5,6 +5,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.sql.Timestamp;
 
 @Getter
@@ -15,6 +16,6 @@ public class Account {
     private Long id;
     private String publicId;
     private String name;
-    private double balance;
+    private BigDecimal balance;
     private Timestamp created;
 }
